+++ components/generic/Toaster.jsx
@@ -6,7 +6,7 @@ const Toaster = (props) => {
   return (
     <div
       id="toast-success"
-      className={`absolute bottom-5 right-5 flex items-center w-full max-w-xs p-4 mb-4 text-gray-500 bg-${color}-200 rounded-lg shadow dark:text-gray-400 dark:bg-gray-800`}
+      className={`absolute bottom-5 right-5 flex items-center w-full max-w-xs p-4 mb-4 text-gray-500 bg-${color}-200 rounded-lg shadow dark:text-gray-400 dark:bg-gray-800 z-10`}
       role="alert"
     >
       <div
